style: Revert layout to #44's state
