+++ recipe-app/src/CSS/App.css
@@ -1,12 +1,9 @@
 #root {
   width: 100vw;
-  display: flex;
-  flex-wrap: wrap;
 }
 
 .page {
   min-height: 100vh;
-  flex-grow: 1;
 }
 
 /***********/
