+++ recipe-app/src/CSS/Gallery.css
@@ -5,5 +5,4 @@
   grid-template-columns: repeat(auto-fit, minmax(240px, 1fr));
   padding: var(--page-gallery-block-padding) var(--page-gallery-inline-padding);
   background-color: var(--display-background-color-light);
-  width: inherit;
 }
