+++ recipe-app/src/CSS/SideBar.css
@@ -1,5 +1,5 @@
 .mobile__nav {
-  background-color: #303030;
+  background-color: #171717;
   display: flex;
   align-items: center;
   justify-content: space-between;
@@ -18,8 +18,8 @@
   display: none;
   width: var(--page-desktop-aside-width);
   flex-shrink: 0;
-  background-color: #303030;
-  color: white;
+  background-color: #171717;
+  color: #fff;
 }
 
 .container-input {
@@ -37,18 +37,10 @@
 }
 
 .search-box {
-  /*
-  
-  border: 1px solid black;
-  border-right: none;
-  outline: none;
-
-   */
   box-shadow: rgba(0, 0, 0, 0.24) 0px 3px 8px;
   border: none;
   border-right: none;
-    height: 35px;
-
+  height: 35px;
   outline: none;
   color: black;
   text-align: center;
@@ -56,7 +48,7 @@
   border-top-left-radius: 10px;
   border-bottom-left-radius: 10px;
   padding: 0.5em 0;
-  background: white;
+  background: #fff;
 }
 
 .search-btn {
@@ -134,8 +126,8 @@ input.checkbox {
 
 .nav-mobile {
   padding: 1em 1.5em 2em;
-  background-color: #303030;
-  color: white;
+  background-color: #171717;
+  color: #fff;
 }
 
 @media (min-width: 600px) {
