+++ recipe-app/src/CSS/Welcome.css
@@ -2,32 +2,31 @@
 
 .welcome-modal {
   color: #fff;
-  background-color: black;
+  background-color: #000;
   background-repeat: no-repeat;
   background-position: center;
   background-size: cover;
+  position: fixed;
+  top: 0;
   z-index: 500;
   margin-left: calc(-1 * var(--page-gallery-inline-padding));
+  width: 100vw;
+  height: 100vh;
   animation-name: fade-out-welcome-modal;
   animation-delay: 3s;
   animation-duration: 4.5s;
   animation-timing-function: ease-in;
   animation-fill-mode: forwards;
-  display: flex;
-  /* Activate display: none to temporarily disable the welcome modal, e.g., while testing the app */
-  /* display: none; */
-  position: fixed;
-  left: calc(var(--page-desktop-aside-width) + 4em);
-  height: 100%;
 }
 
 .welcome-modal__overlay {
   box-sizing: border-box;
   display: flex;
-  flex-grow: 1;
   align-items: center;
   justify-content: center;
   background-color: rgba(0, 0, 0, 0.33);
+  width: inherit;
+  height: inherit;
 }
 
 .welcome-modal__heading {
@@ -47,9 +46,9 @@
   }
 }
 
-/*********************/
-/* Tablet and Desktop*/
-/*********************/
+/**********************/
+/* Tablet and Desktop */
+/**********************/
 
 @media (min-width: 600px) {
   .welcome-modal__heading {
