+++ recipe-app/src/CSS/index.css
@@ -4,8 +4,10 @@
   font-weight: 400;
 
   color-scheme: light dark;
+  color: rgba(255, 255, 255, 0.87);
+  background-color: #242424;
   --font-color-light: #3f3e3e;
-  --card-font-color-light: #ECECEC;
+  --card-font-color-light: #ececec;
   --sidebar-background-color-light: #40a969;
   --display-background-color-light: #d6d2c3;
 
@@ -37,6 +39,7 @@ body {
   min-width: 320px;
   min-height: 100vh;
 }
+
 h1 {
   font-size: 3.2em;
   line-height: 1.1;
